fix(planning): removed things used for testing and adapted unit testing to the new cone coloring

+++ src/planning/src/planning/path_calculation.cpp
@@ -60,7 +60,8 @@ std::vector<PathPoint> PathCalculation::process_delaunay_triangulations(
     }
   }
 
-  RCLCPP_DEBUG(rclcpp::get_logger("rclcpp"), "Unordered path size: %d", unordered_path.size());
+  RCLCPP_DEBUG(rclcpp::get_logger("rclcpp"), "Unordered path size: %d",
+               static_cast<int>(unordered_path.size()));
 
   return unordered_path;
 }
